[IOTDB-6084] Pipe: Support node-urls (client HA) in iotdb-thrift-connector-v1(v2)
